Extract controller from main file/introduce dataclasses
